+++ packages/react-container-query/src/index.js
@@ -1,35 +1,8 @@
-import React, { Component } from "react";
-import ReactDOM from "react-dom";
-import Container from "@zeecoder/container-query";
+import ContainerQuery from "./ContainerQuery";
+import ResizeObserver from "./ResizeObserver";
+import withContainerQuery from "./withContainerQuery";
 
-/**
- * A HoC to initialise a single `Container` based on the given stats.
- *
- * @param {Component} WrappedComponent
- * @param {{}} stats
- * @param {{}} [options]
- * @return {Component}
- */
-const withContainerQuery = (WrappedComponent, stats, options = {}) => {
-  const HoC = class extends Component {
-    componentDidMount() {
-      new Container(ReactDOM.findDOMNode(this), stats, options);
-    }
-
-    render() {
-      const { wrappedComponentRef, ...wrappedProps } = this.props;
-
-      return <WrappedComponent ref={wrappedComponentRef} {...wrappedProps} />;
-    }
-  };
-
-  HoC.displayName =
-    "withContainerQuery(" +
-    (WrappedComponent.displayName || WrappedComponent.name) +
-    ")";
-  HoC.WrappedComponent = WrappedComponent;
-
-  return HoC;
-};
-
+export { ContainerQuery };
+export { ResizeObserver };
+export { withContainerQuery };
 export default withContainerQuery;